+++ concordium-node/src/dumper.rs
@@ -1,6 +1,6 @@
 //! Handles the `network_dump` feature.
 
-cfg_if! {
+cfg_if::cfg_if! {
     if #[cfg(feature = "network_dump")] {
         use crate::common::P2PNodeId;
         use crossbeam_channel::{self, Receiver};