+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Added command-line option for specifying a timeout when downloading
   block indices and catch-up chunk files specified by an URL. The default
   timeout is 5 now minutes per chunk instead of waiting indefinitely.
+- Remove the "instrumentation" feature of the node and build the node with
+  Prometheus support enabled by default.
+  - Remove the `CONCORDIUM_NODE_PROMETHEUS_SERVER` environment variable.
+    The prometheus server is now started if
+    `CONCORDIUM_NODE_PROMETHEUS_LISTEN_PORT` is set.
 
 ## 5.1.1